+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 ### Changed
 - The `data` attribute can now be updated without specifying a `value`.
 - The attribute type can now be set individually using the expert mode.
+- Removed deprecated syntax for changing attributes
 
 ## [0.7.1] - 2020-09-28
 ### Changed